Cleanup leftovers frmo API Connexion
