+++ airflow/cli/commands/remote_commands/dag_command.py
@@ -29,16 +29,19 @@ import subprocess
 import sys
 from typing import TYPE_CHECKING
 
+from itsdangerous import URLSafeSerializer
+from marshmallow import fields
+from marshmallow_sqlalchemy import SQLAlchemySchema, auto_field
 from sqlalchemy import func, select
 
 from airflow.api.client import get_current_api_client
-from airflow.api_connexion.schemas.dag_schema import dag_schema
 from airflow.cli.simple_table import AirflowConsole
 from airflow.cli.utils import fetch_dag_run_from_run_id_or_logical_date_string
+from airflow.configuration import conf
 from airflow.dag_processing.bundles.manager import DagBundlesManager
 from airflow.exceptions import AirflowException
 from airflow.jobs.job import Job
-from airflow.models import DagBag, DagModel, DagRun, TaskInstance
+from airflow.models import DagBag, DagModel, DagRun, DagTag, TaskInstance
 from airflow.models.errors import ParseImportError
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.sdk.definitions._internal.dag_parsing_context import _airflow_parsing_context_manager
@@ -57,6 +60,66 @@ if TYPE_CHECKING:
     from airflow.models.dag import DAG
     from airflow.timetables.base import DataInterval
 log = logging.getLogger(__name__)
+
+
+# Note: To clean up api_connexion, we need to move the below 2 classes to this file until migrated to FastAPI
+class DagTagSchema(SQLAlchemySchema):
+    """Dag Tag schema."""
+
+    class Meta:
+        """Meta."""
+
+        model = DagTag
+
+    name = auto_field()
+
+
+class DAGSchema(SQLAlchemySchema):
+    """DAG schema."""
+
+    class Meta:
+        """Meta."""
+
+        model = DagModel
+
+    dag_id = auto_field(dump_only=True)
+    dag_display_name = fields.String(attribute="dag_display_name", dump_only=True)
+    bundle_name = auto_field(dump_only=True)
+    bundle_version = auto_field(dump_only=True)
+    is_paused = auto_field()
+    is_active = auto_field(dump_only=True)
+    last_parsed_time = auto_field(dump_only=True)
+    last_expired = auto_field(dump_only=True)
+    default_view = auto_field(dump_only=True)
+    fileloc = auto_field(dump_only=True)
+    file_token = fields.Method("get_token", dump_only=True)
+    owners = fields.Method("get_owners", dump_only=True)
+    description = auto_field(dump_only=True)
+    timetable_summary = auto_field(dump_only=True)
+    timetable_description = auto_field(dump_only=True)
+    tags = fields.List(fields.Nested(DagTagSchema), dump_only=True)
+    max_active_tasks = auto_field(dump_only=True)
+    max_active_runs = auto_field(dump_only=True)
+    max_consecutive_failed_dag_runs = auto_field(dump_only=True)
+    has_task_concurrency_limits = auto_field(dump_only=True)
+    has_import_errors = auto_field(dump_only=True)
+    next_dagrun = auto_field(dump_only=True)
+    next_dagrun_data_interval_start = auto_field(dump_only=True)
+    next_dagrun_data_interval_end = auto_field(dump_only=True)
+    next_dagrun_create_after = auto_field(dump_only=True)
+
+    @staticmethod
+    def get_owners(obj: DagModel):
+        """Convert owners attribute to DAG representation."""
+        if not getattr(obj, "owners", None):
+            return []
+        return obj.owners.split(",")
+
+    @staticmethod
+    def get_token(obj: DagModel):
+        """Return file token."""
+        serializer = URLSafeSerializer(conf.get_mandatory_value("webserver", "secret_key"))
+        return serializer.dumps(obj.fileloc)
 
 
 @cli_utils.action_cli
@@ -329,15 +392,16 @@ def dag_next_execution(args) -> None:
 def dag_list_dags(args, session: Session = NEW_SESSION) -> None:
     """Display dags with or without stats at the command line."""
     cols = args.columns if args.columns else []
-    invalid_cols = [c for c in cols if c not in dag_schema.fields]
-    valid_cols = [c for c in cols if c in dag_schema.fields]
+    dag_schema_fields = DAGSchema().fields
+    invalid_cols = [c for c in cols if c not in dag_schema_fields]
+    valid_cols = [c for c in cols if c in dag_schema_fields]
 
     if invalid_cols:
         from rich import print as rich_print
 
         rich_print(
             f"[red][bold]Error:[/bold] Ignoring the following invalid columns: {invalid_cols}.  "
-            f"List of valid columns: {list(dag_schema.fields.keys())}",
+            f"List of valid columns: {list(dag_schema_fields.keys())}",
             file=sys.stderr,
         )
 
@@ -363,7 +427,7 @@ def dag_list_dags(args, session: Session = NEW_SESSION) -> None:
     def get_dag_detail(dag: DAG) -> dict:
         dag_model = DagModel.get_dagmodel(dag.dag_id, session=session)
         if dag_model:
-            dag_detail = dag_schema.dump(dag_model)
+            dag_detail = DAGSchema().dump(dag_model)
         else:
             dag_detail = _get_dagbag_dag_details(dag)
         return {col: dag_detail[col] for col in valid_cols}
@@ -395,7 +459,7 @@ def dag_details(args, session: Session = NEW_SESSION):
     dag = DagModel.get_dagmodel(args.dag_id, session=session)
     if not dag:
         raise SystemExit(f"DAG: {args.dag_id} does not exist in 'dag' table")
-    dag_detail = dag_schema.dump(dag)
+    dag_detail = DAGSchema().dump(dag)
 
     if args.output in ["table", "plain"]:
         data = [{"property_name": key, "property_value": value} for key, value in dag_detail.items()]
